Add a dependency on pkgload.
orderly has a bit of magic to load a transitional `orderly2` package if
it detects the use of it in a report. That functionality is implemented
using pkgload, but pkgload was never added as an `Imports`.

Most people will have pkgload installed anyway, but the Docker image for
the orderly.runner does not, meaning the compatibility code actually
fails on Packit.

If we add it as a proper `Imports` here it will get pulled on the next
runner image build.